requirements.txt
requests
python-dotenv
langchain~=0.3.25
langchain-openai
openai
fastapi~=0.115.12
uvicorn
requests
python-dotenv
langchain
langchain-openai
openai
pydantic

langchain-community~=0.3.24